code reformatted to incorporate new libraries. Code still uses legacy methods for buttonmaking and doesnt take advantage of the features of TFT_eSPI. However, this was supposed to be a quick and dirty upgrade
No known bugs, settings toggle and save. Graphics are sligned.

TODO: Incorporate auto calibrate feature on the devices first run. Have the device check if the settings have been created and saved before, if not, then run the calibration script. After the values are saved, have it save those constants for the rest of the program to use then continue normally